+++ test/test_library.py
@@ -449,6 +449,16 @@ class DestinationTest(_common.TestCase):
         self.assertEqual(self.i.destination(),
                          np('base/ber/foo'))
 
+    def test_destination_with_replacements_argument(self):
+        self.lib.directory = b'base'
+        self.lib.replacements = [(re.compile(r'a'), 'f')]
+        self.lib.path_formats = [('default', '$album/$title')]
+        self.i.title = 'foo'
+        self.i.album = 'bar'
+        replacements = [(re.compile(r'a'), 'e')]
+        self.assertEqual(self.i.destination(replacements=replacements),
+                         np('base/ber/foo'))
+
     @unittest.skip('unimplemented: #359')
     def test_destination_with_empty_component(self):
         self.lib.directory = b'base'